+++ main.tf
@@ -1,32 +1,44 @@
 locals {
-  port            = var.port == "" ? var.engine == "aurora-postgresql" ? "5432" : "3306" : var.port
-  master_password = var.password == "" ? random_id.master_password.b64 : var.password
+  port                 = var.port == "" ? var.engine == "aurora-postgresql" ? "5432" : "3306" : var.port
+  master_password      = var.password == "" ? random_password.master_password.result : var.password
+  db_subnet_group_name = var.db_subnet_group_name == "" ? join("", aws_db_subnet_group.this.*.name) : var.db_subnet_group_name
+  backtrack_window     = (var.engine == "aurora-mysql" || var.engine == "aurora") && var.engine_mode != "serverless" ? var.backtrack_window : 0
+
+  rds_enhanced_monitoring_arn  = join("", aws_iam_role.rds_enhanced_monitoring.*.arn)
+  rds_enhanced_monitoring_name = join("", aws_iam_role.rds_enhanced_monitoring.*.name)
+
+  rds_security_group_id = join("", aws_security_group.this.*.id)
+
+  name = "aurora-${var.name}"
 }
 
 # Random string to use as master password unless one is specified
-resource "random_id" "master_password" {
-  byte_length = 10
+resource "random_password" "master_password" {
+  length  = 10
+  special = false
 }
 
 resource "aws_db_subnet_group" "this" {
+  count = var.db_subnet_group_name == "" ? 1 : 0
+
   name        = var.name
   description = "For Aurora cluster ${var.name}"
   subnet_ids  = var.subnets
 
-  tags = merge(
-    var.tags,
-    {
-      "Name" = "aurora-${var.name}"
-    },
-  )
+  tags = merge(var.tags, {
+    Name = local.name
+  })
 }
 
 resource "aws_rds_cluster" "this" {
   global_cluster_identifier           = var.global_cluster_identifier
   cluster_identifier                  = var.name
+  replication_source_identifier       = var.replication_source_identifier
+  source_region                       = var.source_region
   engine                              = var.engine
   engine_mode                         = var.engine_mode
   engine_version                      = var.engine_version
+  enable_http_endpoint                = var.enable_http_endpoint
   kms_key_id                          = var.kms_key_id
   database_name                       = var.database_name
   master_username                     = var.username
@@ -38,15 +50,31 @@ resource "aws_rds_cluster" "this" {
   preferred_backup_window             = var.preferred_backup_window
   preferred_maintenance_window        = var.preferred_maintenance_window
   port                                = local.port
-  db_subnet_group_name                = aws_db_subnet_group.this.name
-  vpc_security_group_ids              = concat([aws_security_group.this.id], var.vpc_security_group_ids)
+  db_subnet_group_name                = local.db_subnet_group_name
+  vpc_security_group_ids              = compact(concat(aws_security_group.this.*.id, var.vpc_security_group_ids))
   snapshot_identifier                 = var.snapshot_identifier
   storage_encrypted                   = var.storage_encrypted
   apply_immediately                   = var.apply_immediately
   db_cluster_parameter_group_name     = var.db_cluster_parameter_group_name
   iam_database_authentication_enabled = var.iam_database_authentication_enabled
+  backtrack_window                    = local.backtrack_window
+  copy_tags_to_snapshot               = var.copy_tags_to_snapshot
+  iam_roles                           = var.iam_roles
 
   enabled_cloudwatch_logs_exports = var.enabled_cloudwatch_logs_exports
+
+  dynamic "scaling_configuration" {
+    for_each = length(keys(var.scaling_configuration)) == 0 ? [] : [
+    var.scaling_configuration]
+
+    content {
+      auto_pause               = lookup(scaling_configuration.value, "auto_pause", null)
+      max_capacity             = lookup(scaling_configuration.value, "max_capacity", null)
+      min_capacity             = lookup(scaling_configuration.value, "min_capacity", null)
+      seconds_until_auto_pause = lookup(scaling_configuration.value, "seconds_until_auto_pause", null)
+      timeout_action           = lookup(scaling_configuration.value, "timeout_action", null)
+    }
+  }
 
   tags = var.tags
 }
@@ -60,16 +88,17 @@ resource "aws_rds_cluster_instance" "this" {
   engine_version                  = var.engine_version
   instance_class                  = count.index > 0 ? coalesce(var.instance_type_replica, var.instance_type) : var.instance_type
   publicly_accessible             = var.publicly_accessible
-  db_subnet_group_name            = aws_db_subnet_group.this.name
+  db_subnet_group_name            = local.db_subnet_group_name
   db_parameter_group_name         = var.db_parameter_group_name
   preferred_maintenance_window    = var.preferred_maintenance_window
   apply_immediately               = var.apply_immediately
-  monitoring_role_arn             = join("", aws_iam_role.rds_enhanced_monitoring.*.arn)
+  monitoring_role_arn             = local.rds_enhanced_monitoring_arn
   monitoring_interval             = var.monitoring_interval
   auto_minor_version_upgrade      = var.auto_minor_version_upgrade
   promotion_tier                  = count.index + 1
   performance_insights_enabled    = var.performance_insights_enabled
   performance_insights_kms_key_id = var.performance_insights_kms_key_id
+  ca_cert_identifier              = var.ca_cert_identifier
 
   tags = var.tags
 }
@@ -98,12 +127,18 @@ resource "aws_iam_role" "rds_enhanced_monitoring" {
 
   name               = "rds-enhanced-monitoring-${var.name}"
   assume_role_policy = data.aws_iam_policy_document.monitoring_rds_assume_role.json
+
+  permissions_boundary = var.permissions_boundary
+
+  tags = merge(var.tags, {
+    Name = local.name
+  })
 }
 
 resource "aws_iam_role_policy_attachment" "rds_enhanced_monitoring" {
   count = var.monitoring_interval > 0 ? 1 : 0
 
-  role       = aws_iam_role.rds_enhanced_monitoring[0].name
+  role       = local.rds_enhanced_monitoring_name
   policy_arn = "arn:aws:iam::aws:policy/service-role/AmazonRDSEnhancedMonitoringRole"
 }
 
@@ -128,31 +163,52 @@ resource "aws_appautoscaling_policy" "autoscaling_read_replica_count" {
 
   target_tracking_scaling_policy_configuration {
     predefined_metric_specification {
-      predefined_metric_type = "RDSReaderAverageCPUUtilization"
+      predefined_metric_type = var.predefined_metric_type
     }
 
     scale_in_cooldown  = var.replica_scale_in_cooldown
     scale_out_cooldown = var.replica_scale_out_cooldown
-    target_value       = var.replica_scale_cpu
+    target_value       = var.predefined_metric_type == "RDSReaderAverageCPUUtilization" ? var.replica_scale_cpu : var.replica_scale_connections
   }
 
   depends_on = [aws_appautoscaling_target.read_replica_count]
 }
 
 resource "aws_security_group" "this" {
+  count = var.create_security_group ? 1 : 0
+
   name_prefix = "${var.name}-"
   vpc_id      = var.vpc_id
 
-  tags = var.tags
+  description = var.security_group_description == "" ? "Control traffic to/from RDS Aurora ${var.name}" : var.security_group_description
+
+  tags = merge(var.tags, {
+    Name = local.name
+  })
 }
 
 resource "aws_security_group_rule" "default_ingress" {
-  count = var.allowed_security_groups_count
+  count = var.create_security_group ? length(var.allowed_security_groups) : 0
+
+  description = "From allowed SGs"
 
   type                     = "ingress"
   from_port                = aws_rds_cluster.this.port
   to_port                  = aws_rds_cluster.this.port
   protocol                 = "tcp"
   source_security_group_id = element(var.allowed_security_groups, count.index)
-  security_group_id        = aws_security_group.this.id
+  security_group_id        = local.rds_security_group_id
+}
+
+resource "aws_security_group_rule" "cidr_ingress" {
+  count = var.create_security_group && length(var.allowed_cidr_blocks) > 0 ? 1 : 0
+
+  description = "From allowed CIDRs"
+
+  type              = "ingress"
+  from_port         = aws_rds_cluster.this.port
+  to_port           = aws_rds_cluster.this.port
+  protocol          = "tcp"
+  cidr_blocks       = var.allowed_cidr_blocks
+  security_group_id = local.rds_security_group_id
 }