+++ variables.tf
@@ -1,3 +1,9 @@
+variable "create_security_group" {
+  description = "Whether to create security group for RDS cluster"
+  type        = bool
+  default     = true
+}
+
 variable "name" {
   description = "Name given resources"
   type        = string
@@ -6,6 +12,7 @@ variable "name" {
 variable "subnets" {
   description = "List of subnet IDs to use"
   type        = list(string)
+  default     = []
 }
 
 variable "replica_count" {
@@ -15,12 +22,14 @@ variable "replica_count" {
 
 variable "allowed_security_groups" {
   description = "A list of Security Group ID's to allow access to."
+  type        = list(string)
   default     = []
 }
 
-variable "allowed_security_groups_count" {
-  description = "The number of Security Groups being added, terraform doesn't let us use length() in a count field"
-  default     = 0
+variable "allowed_cidr_blocks" {
+  description = "A list of CIDR blocks which are allowed to access the database"
+  type        = list(string)
+  default     = []
 }
 
 variable "vpc_id" {
@@ -126,13 +135,19 @@ variable "auto_minor_version_upgrade" {
 variable "db_parameter_group_name" {
   description = "The name of a DB parameter group to use"
   type        = string
-  default     = "default.aurora5.6"
+  default     = null
 }
 
 variable "db_cluster_parameter_group_name" {
   description = "The name of a DB Cluster parameter group to use"
   type        = string
-  default     = "default.aurora5.6"
+  default     = null
+}
+
+variable "scaling_configuration" {
+  description = "Map of nested attributes with scaling properties. Only valid when engine_mode is set to `serverless`"
+  type        = map(string)
+  default     = {}
 }
 
 variable "snapshot_identifier" {
@@ -165,6 +180,12 @@ variable "engine_version" {
   default     = "5.6.10a"
 }
 
+variable "enable_http_endpoint" {
+  description = "Whether or not to enable the Data API for a serverless Aurora database engine."
+  type        = bool
+  default     = false
+}
+
 variable "replica_scale_enabled" {
   description = "Whether to enable autoscaling for RDS Aurora (MySQL) read replicas"
   type        = bool
@@ -187,6 +208,12 @@ variable "replica_scale_cpu" {
   description = "CPU usage to trigger autoscaling at"
   type        = number
   default     = 70
+}
+
+variable "replica_scale_connections" {
+  description = "Average number of connections to trigger autoscaling at. Default value is 70% of db.r4.large's default max_connections"
+  type        = number
+  default     = 700
 }
 
 variable "replica_scale_in_cooldown" {
@@ -238,13 +265,70 @@ variable "global_cluster_identifier" {
 }
 
 variable "engine_mode" {
-  description = "The database engine mode. Valid values: global, parallelquery, provisioned, serverless."
+  description = "The database engine mode. Valid values: global, parallelquery, provisioned, serverless, multimaster."
   type        = string
   default     = "provisioned"
 }
 
+variable "replication_source_identifier" {
+  description = "ARN of a source DB cluster or DB instance if this DB cluster is to be created as a Read Replica."
+  default     = ""
+}
+
+variable "source_region" {
+  description = "The source region for an encrypted replica DB cluster."
+  default     = ""
+}
+
 variable "vpc_security_group_ids" {
   description = "List of VPC security groups to associate to the cluster in addition to the SG we create in this module"
-  type        = list
+  type        = list(string)
   default     = []
+}
+
+variable "db_subnet_group_name" {
+  description = "The existing subnet group name to use"
+  type        = string
+  default     = ""
+}
+
+variable "predefined_metric_type" {
+  description = "The metric type to scale on. Valid values are RDSReaderAverageCPUUtilization and RDSReaderAverageDatabaseConnections."
+  default     = "RDSReaderAverageCPUUtilization"
+}
+
+variable "backtrack_window" {
+  description = "The target backtrack window, in seconds. Only available for aurora engine currently. To disable backtracking, set this value to 0. Defaults to 0. Must be between 0 and 259200 (72 hours)"
+  type        = number
+  default     = 0
+}
+
+variable "copy_tags_to_snapshot" {
+  description = "Copy all Cluster tags to snapshots."
+  type        = bool
+  default     = false
+}
+
+variable "iam_roles" {
+  description = "A List of ARNs for the IAM roles to associate to the RDS Cluster."
+  type        = list(string)
+  default     = []
+}
+
+variable "security_group_description" {
+  description = "The description of the security group. If value is set to empty string it will contain cluster name in the description."
+  type        = string
+  default     = "Managed by Terraform"
+}
+
+variable "permissions_boundary" {
+  description = "The ARN of the policy that is used to set the permissions boundary for the role."
+  type        = string
+  default     = null
+}
+
+variable "ca_cert_identifier" {
+  description = "The identifier of the CA certificate for the DB instance"
+  type        = string
+  default     = "rds-ca-2019"
 }