+++ CHANGELOG.md
@@ -1,20 +1,267 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+
 <a name="unreleased"></a>
 ## [Unreleased]
 
 
 
-<a name="v2.2.1"></a>
-## [v2.2.1] - 2019-08-09
+<a name="v2.22.0"></a>
+## [v2.22.0] - 2020-08-19
 
-- Support for different instance types for master and replica
+- feat: Add output aws_rds_cluster.this.hosted_zone_id ([#125](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/125))
+
+
+<a name="v2.21.0"></a>
+## [v2.21.0] - 2020-08-13
+
+- Updated version requirements for AWS provider v3
+
+
+<a name="v2.20.0"></a>
+## [v2.20.0] - 2020-07-20
+
+- feat: Add tags to RDS enhanced monitoring IAM role ([#136](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/136))
+- Upgraded Terraform version supported
+
+
+<a name="v2.19.0"></a>
+## [v2.19.0] - 2020-06-12
+
+- feat: add "this_rds_cluster_instance_ids" to module outputs ([#107](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/107))
+- docs: Add "multimaster" to engine_mode docs ([#133](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/133))
+
+
+<a name="v2.18.0"></a>
+## [v2.18.0] - 2020-06-10
+
+- feat: Add permissions boundary for IAM role ([#104](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/104))
+- Updated README
+
+
+<a name="v2.17.0"></a>
+## [v2.17.0] - 2020-04-11
+
+- feat: Support enable_http_endpoint - Data API ([#117](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/117))
+- Updated postgres example to latest engine_version
+
+
+<a name="v2.16.0"></a>
+## [v2.16.0] - 2020-02-26
+
+- Updated pre-commit-terraform with README
+- Updated pre-commit-terraform with README
+- Adding support for ca_cert_identifier for the upcoming RDS cert update on 5 March 2020 ([#91](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/91)) ([#98](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/98))
+- Added missing type type for allowed_security_groups variable ([#97](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/97))
+
+
+<a name="v2.15.0"></a>
+## [v2.15.0] - 2020-01-28
+
+- Disabled special characters in random_password ([#101](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/101))
+
+
+<a name="v2.14.0"></a>
+## [v2.14.0] - 2020-01-23
+
+- Changed random_id to random_password resource ([#99](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/99))
+
+
+<a name="v2.13.0"></a>
+## [v2.13.0] - 2020-01-23
+
+- Adding version requirements
+
+
+<a name="v2.12.0"></a>
+## [v2.12.0] - 2020-01-23
+
+- Updated pre-commit-terraform to support terraform-docs 0.8
+- Hide sensitive password data from cli output ([#89](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/89))
+
+
+<a name="v2.11.0"></a>
+## [v2.11.0] - 2019-11-29
+
+- Made description of security group backward compatible (and optional) ([#87](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/87))
+- Add security group name and description ([#80](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/80))
+- Updated README
+- Set null as default db_parameter_group_name ([#72](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/72))
+- Support setting aws_rds_cluster iam roles ([#74](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/74))
+- Fix the error when "create_security_group" is false ([#75](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/75))
+- Added support for conditional creation of SG and allowed CIDR blocks ([#71](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/71))
+- Updated README
+- Add option to customise predefined_metric_type ([#66](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/66))
+- Added support for backtrack_window, cross region read replicas, copy_tags_to_snapshots, scaling_configurations ([#70](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/70))
+- Fixes after [#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64) (subnets are optional now), updated pre-commit hooks and docs
+- Add support to use existing subnet group name ([#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64))
+- Fix PostgreSQL example link ([#52](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/52))
+- Support using previously created SG for Aurora cluster - TF 0.12 ([#49](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/49))
+- Add cluster ARN output as 'this_rds_cluster_arn' ([#48](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/48))
+- Upgraded module to support Terraform 0.12 ([#45](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/45))
+
+### 
+
+on ../../modules/aws-rds-aurora/main.tf line 4, in locals:
+
+4:   db_subnet_group_name = var.db_subnet_group_name == "" ?
+aws_db_subnet_group.this[0].name : var.db_subnet_group_name
+
+|----------------
+
+| aws_db_subnet_group.this is empty tuple
+
+when calling import with this module in the configuration.
+
+
+<a name="v1.21.0"></a>
+## [v1.21.0] - 2019-11-28
+
+- Revert "add support for ca_cert_identifier option ([#83](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/83))" ([#85](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/85))
+- add support for ca_cert_identifier option ([#83](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/83))
+- Updated docs
+- Add scaling_configuration ([#63](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/63))
+- Added ability to deploy a cross region Aurora Cluster Read Replica ([#69](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/69))
+- Added backtrack_windows for Aurora ([#57](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/57))
+- Added changelog
+- Added copy_tags_to_snapshot (closes [#60](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/60))
+- Allow the addition of IP-based ingress rules ([#51](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/51))
+- vpc_security_group_ids for Terraform 0.11 ([#46](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/46))
+
+
+<a name="v2.10.0"></a>
+## [v2.10.0] - 2019-11-28
+
+- Add security group name and description ([#80](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/80))
+- Updated README
+- Set null as default db_parameter_group_name ([#72](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/72))
+- Support setting aws_rds_cluster iam roles ([#74](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/74))
+- Fix the error when "create_security_group" is false ([#75](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/75))
+- Added support for conditional creation of SG and allowed CIDR blocks ([#71](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/71))
+- Updated README
+- Add option to customise predefined_metric_type ([#66](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/66))
+- Added support for backtrack_window, cross region read replicas, copy_tags_to_snapshots, scaling_configurations ([#70](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/70))
+- Fixes after [#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64) (subnets are optional now), updated pre-commit hooks and docs
+- Add support to use existing subnet group name ([#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64))
+- Fix PostgreSQL example link ([#52](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/52))
+- Support using previously created SG for Aurora cluster - TF 0.12 ([#49](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/49))
+- Add cluster ARN output as 'this_rds_cluster_arn' ([#48](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/48))
+- Upgraded module to support Terraform 0.12 ([#45](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/45))
+
+### 
+
+on ../../modules/aws-rds-aurora/main.tf line 4, in locals:
+
+4:   db_subnet_group_name = var.db_subnet_group_name == "" ?
+aws_db_subnet_group.this[0].name : var.db_subnet_group_name
+
+|----------------
+
+| aws_db_subnet_group.this is empty tuple
+
+when calling import with this module in the configuration.
+
+
+<a name="v1.20.0"></a>
+## [v1.20.0] - 2019-11-28
+
+- add support for ca_cert_identifier option ([#83](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/83))
+- Updated docs
+- Add scaling_configuration ([#63](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/63))
+- Added ability to deploy a cross region Aurora Cluster Read Replica ([#69](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/69))
+- Added backtrack_windows for Aurora ([#57](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/57))
+- Added changelog
+- Added copy_tags_to_snapshot (closes [#60](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/60))
+- Allow the addition of IP-based ingress rules ([#51](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/51))
+- vpc_security_group_ids for Terraform 0.11 ([#46](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/46))
+
+
+<a name="v2.9.0"></a>
+## [v2.9.0] - 2019-11-08
+
+- Updated README
+- Set null as default db_parameter_group_name ([#72](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/72))
+
+
+<a name="v2.8.0"></a>
+## [v2.8.0] - 2019-11-08
+
+- Support setting aws_rds_cluster iam roles ([#74](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/74))
+
+
+<a name="v2.7.0"></a>
+## [v2.7.0] - 2019-11-08
+
+- Fix the error when "create_security_group" is false ([#75](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/75))
+
+
+<a name="v2.6.0"></a>
+## [v2.6.0] - 2019-09-30
+
+- Added support for conditional creation of SG and allowed CIDR blocks ([#71](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/71))
+
+
+<a name="v2.5.0"></a>
+## [v2.5.0] - 2019-09-30
+
+- Updated README
+- Add option to customise predefined_metric_type ([#66](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/66))
+
+
+<a name="v2.4.0"></a>
+## [v2.4.0] - 2019-09-30
+
+- Added support for backtrack_window, cross region read replicas, copy_tags_to_snapshots, scaling_configurations ([#70](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/70))
+- Fixes after [#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64) (subnets are optional now), updated pre-commit hooks and docs
+- Add support to use existing subnet group name ([#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64))
+- Fix PostgreSQL example link ([#52](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/52))
+- Support using previously created SG for Aurora cluster - TF 0.12 ([#49](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/49))
+- Add cluster ARN output as 'this_rds_cluster_arn' ([#48](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/48))
+- Upgraded module to support Terraform 0.12 ([#45](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/45))
+
+
+<a name="v1.19.0"></a>
+## [v1.19.0] - 2019-09-29
+
+- Updated docs
+- Add scaling_configuration ([#63](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/63))
+
+
+<a name="v1.18.0"></a>
+## [v1.18.0] - 2019-09-29
+
+- Added ability to deploy a cross region Aurora Cluster Read Replica ([#69](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/69))
+
+
+<a name="v1.17.0"></a>
+## [v1.17.0] - 2019-09-28
+
+- Added backtrack_windows for Aurora ([#57](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/57))
+
+
+<a name="v1.16.0"></a>
+## [v1.16.0] - 2019-09-28
+
+- Added changelog
+- Added copy_tags_to_snapshot (closes [#60](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/60))
+- Allow the addition of IP-based ingress rules ([#51](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/51))
+- vpc_security_group_ids for Terraform 0.11 ([#46](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/46))
+
+
+<a name="v2.3.0"></a>
+## [v2.3.0] - 2019-08-29
+
+- Fixes after [#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64) (subnets are optional now), updated pre-commit hooks and docs
+- Add support to use existing subnet group name ([#64](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/64))
+- Fix PostgreSQL example link ([#52](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/52))
+
 
 <a name="v2.2.0"></a>
 ## [v2.2.0] - 2019-06-24
 
 - Support using previously created SG for Aurora cluster - TF 0.12 ([#49](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/49))
-- Updated CHANGELOG
 - Add cluster ARN output as 'this_rds_cluster_arn' ([#48](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/48))
-- Updated CHANGELOG
 - Upgraded module to support Terraform 0.12 ([#45](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/45))
 
 
@@ -28,9 +275,7 @@
 <a name="v2.1.0"></a>
 ## [v2.1.0] - 2019-06-14
 
-- Updated CHANGELOG
 - Add cluster ARN output as 'this_rds_cluster_arn' ([#48](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/48))
-- Updated CHANGELOG
 - Upgraded module to support Terraform 0.12 ([#45](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/45))
 
 
@@ -43,7 +288,6 @@
 <a name="v2.0.0"></a>
 ## [v2.0.0] - 2019-06-11
 
-- Updated CHANGELOG
 - Upgraded module to support Terraform 0.12 ([#45](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/45))
 
 
@@ -79,7 +323,6 @@
 <a name="v1.9.0"></a>
 ## [v1.9.0] - 2019-02-16
 
-- Merge pull request [#25](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/25) from Blokje5/add-support-for-database-iam-authentication
 - Remove validation
 - Add support for IAM Database authentication
 
@@ -89,21 +332,18 @@
 
 - Updated example with correct allowed_security_groups_count
 - Run pre-commit hooks
-- Merge pull request [#23](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/23) from dekimsey/fix-default-ingress-count-error
 - Hardcode number of allowed_security_groups
 
 
 <a name="v1.7.0"></a>
 ## [v1.7.0] - 2019-01-30
 
-- Merge pull request [#21](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/21) from max-rocket-internet/remove_maintainer
 - Removing myself from README
 
 
 <a name="v1.6.0"></a>
 ## [v1.6.0] - 2019-01-30
 
-- Merge pull request [#22](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/22) from terraform-aws-modules/pr/13
 - Merge remote-tracking branch 'origin/master' into pr-13
 - Added output for database_name
 - add support for database_name argument
@@ -112,7 +352,6 @@
 <a name="v1.5.0"></a>
 ## [v1.5.0] - 2019-01-30
 
-- Merge pull request [#15](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/15) from christiangjengedal/support_deletion_protection
 - Merge branch 'master' into support_deletion_protection
 - Added argument deletion_protection
 
@@ -120,7 +359,6 @@
 <a name="v1.4.0"></a>
 ## [v1.4.0] - 2019-01-30
 
-- Merge pull request [#20](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/20) from terraform-aws-modules/removed_azs
 - Removed availability_zones var (closes [#10](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/10))
 
 
@@ -128,7 +366,6 @@
 ## [v1.3.0] - 2019-01-30
 
 - Run pre-commit hooks
-- Merge pull request [#16](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/16) from terraform-aws-modules/fix_az
 - Merge branch 'master' into fix_az
 - Fixed variable for az
 
@@ -137,7 +374,6 @@
 ## [v1.1.0] - 2019-01-30
 
 - Run pre-commit hooks
-- Merge pull request [#17](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/17) from christiangjengedal/mod_securitygroup
 - 3 adjustments: - Avoid description attr on aws_security_group to avoid forcing new instance on changes - Use name-prefix to ensure unique names - Avoid unnecessary name-prefixes
 
 
@@ -145,12 +381,9 @@
 ## [v1.0.0] - 2018-12-11
 
 - Updated pre-commit hooks version
-- Merge pull request [#4](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/4) from FutureSharks/example_fix
-- Merge pull request [#3](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/3) from terraform-aws-modules/post_initial_commit
 - Fix for: 'count' cannot be computed
 - Added few files
 - Made small changes to be closer to other terraform-aws-modules
-- Merge pull request [#1](https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/1) from FutureSharks/initial_commit
 - adding pre commit hooks config
 - updating readme, docs, main.tf etc
 - adding port output
@@ -167,8 +400,33 @@
 - Initial commit
 
 
-[Unreleased]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.2.1...HEAD
-[v2.2.1]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.2.0...v2.2.1
+[Unreleased]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.22.0...HEAD
+[v2.22.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.21.0...v2.22.0
+[v2.21.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.20.0...v2.21.0
+[v2.20.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.19.0...v2.20.0
+[v2.19.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.18.0...v2.19.0
+[v2.18.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.17.0...v2.18.0
+[v2.17.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.16.0...v2.17.0
+[v2.16.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.15.0...v2.16.0
+[v2.15.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.14.0...v2.15.0
+[v2.14.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.13.0...v2.14.0
+[v2.13.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.12.0...v2.13.0
+[v2.12.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.11.0...v2.12.0
+[v2.11.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.21.0...v2.11.0
+[v1.21.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.10.0...v1.21.0
+[v2.10.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.20.0...v2.10.0
+[v1.20.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.9.0...v1.20.0
+[v2.9.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.8.0...v2.9.0
+[v2.8.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.7.0...v2.8.0
+[v2.7.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.6.0...v2.7.0
+[v2.6.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.5.0...v2.6.0
+[v2.5.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.4.0...v2.5.0
+[v2.4.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.19.0...v2.4.0
+[v1.19.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.18.0...v1.19.0
+[v1.18.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.17.0...v1.18.0
+[v1.17.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.16.0...v1.17.0
+[v1.16.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.3.0...v1.16.0
+[v2.3.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.2.0...v2.3.0
 [v2.2.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.15.0...v2.2.0
 [v1.15.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v2.1.0...v1.15.0
 [v2.1.0]: https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/compare/v1.14.0...v2.1.0